+++ tests/test_constants.py
@@ -5,9 +5,9 @@ import unittest
 from mock import patch
 
 
-from common_python.constants_functions import translate_country_input_to_geofabrik
-from common_python.constants_functions import translate_tags_to_keep
-import common_python.constants as const
+from wahoo_mc.constants_functions import translate_country_input_to_geofabrik
+from wahoo_mc.constants_functions import translate_tags_to_keep
+import wahoo_mc.constants as const
 
 
 class TestTranslateCountries(unittest.TestCase):
