+++ tests/test_geofabrik.py
@@ -6,10 +6,10 @@ import os
 import unittest
 
 # import custom python packages
-from common_python.geofabrik import Geofabrik
-from common_python import file_directory_functions as fd_fct
-from common_python import constants_functions as const_fct
-from common_python.downloader import Downloader
+from wahoo_mc.geofabrik import Geofabrik
+from wahoo_mc import file_directory_functions as fd_fct
+from wahoo_mc import constants_functions as const_fct
+from wahoo_mc.downloader import Downloader
 
 
 def calc_tiles_via_geofabrik_json(input_argument):
