+++ wahoo_mc/__main__.py
@@ -1,7 +1,7 @@
 """
 can be directly called via `python -m common_python -h`
 """
-from common_python import wahoo_map_creator
+from wahoo_mc import wahoo_map_creator
 
 if __name__ == '__main__':
     wahoo_map_creator.run()
