+++ wahoo_mc/downloader.py
@@ -12,8 +12,8 @@ import time
 import logging
 
 # import custom python packages
-from common_python import file_directory_functions as fd_fct
-from common_python import constants_functions as const_fct
+from wahoo_mc import file_directory_functions as fd_fct
+from wahoo_mc import constants_functions as const_fct
 
 log = logging.getLogger('main-logger')
