+++ wahoo_mc/wahoo_map_creator.py
@@ -7,14 +7,15 @@ executable file to create up-to-date map-files for the Wahoo ELEMNT and Wahoo EL
 import logging
 
 # import custom python packages
-from common_python.input import process_call_of_the_tool
-from common_python.file_directory_functions import initialize_work_directories
-from common_python.osm_maps_functions import OsmMaps
+from wahoo_mc.input import process_call_of_the_tool
+from wahoo_mc.file_directory_functions import initialize_work_directories
+from wahoo_mc.osm_maps_functions import OsmMaps
 
 # logging used in the terminal output:
 # # means top-level command
 # ! means error
 # + means additional comment in a working-unit
+
 
 def run():
     # create logger
@@ -37,7 +38,6 @@ def run():
     oOSMmaps.process_input(oInputData.process_border_countries)
     oOSMmaps.check_and_download_files()
 
-
     if oInputData.only_merge is False:
         # Filter tags from country osm.pbf files'
         oOSMmaps.filter_tags_from_country_osm_pbf_files()
@@ -56,7 +56,8 @@ def run():
         oInputData.process_border_countries)
 
     # Creating .map files
-    oOSMmaps.create_map_files(oInputData.save_cruiser, oInputData.tag_wahoo_xml)
+    oOSMmaps.create_map_files(oInputData.save_cruiser,
+                              oInputData.tag_wahoo_xml)
 
     # Zip .map.lzma files
     oOSMmaps.make_and_zip_files(oInputData.keep_map_folders, '.map.lzma')
